# Vectorization of linalg.fill 