--- a/compiler/plugins/target/AMD-AIE/aievec/VectorToVectorConversions.cpp
+++ b/compiler/plugins/target/AMD-AIE/aievec/VectorToVectorConversions.cpp
@@ -17,6 +17,7 @@
 #include "Passes.h"
 #include "aievec/AIEVecOps.h"
 #include "iree-amd-aie/Transforms/Utils/AMDAIEUtils.h"
+#include "iree-amd-aie/aie_runtime/AMDAIEEnums.h"
 #include "iree-amd-aie/aie_runtime/iree_aie_runtime.h"
 #include "llvm/ADT/STLExtras.h"
 #include "mlir/Conversion/AffineToStandard/AffineToStandard.h"
@@ -24,7 +25,11 @@
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/LLVMIR/LLVMAttrs.h"
+#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
+#include "mlir/Dialect/LLVMIR/LLVMTypes.h"
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/Dialect/Utils/ReshapeOpsUtils.h"
 #include "mlir/Dialect/Vector/IR/VectorOps.h"
 #include "mlir/Dialect/Vector/Transforms/LoweringPatterns.h"
@@ -104,7 +109,7 @@ static SmallVector<Value> getCollapsedIndices(RewriterBase &rewriter,
   OpFoldResult collapsedOffset =
       rewriter.create<arith::ConstantIndexOp>(loc, 0).getResult();
 
-  auto collapsedStrides = computeSuffixProduct(
+  SmallVector<int64_t> collapsedStrides = computeSuffixProduct(
       ArrayRef<int64_t>(shape.begin() + firstDimToCollapse, shape.end()));
 
   // Compute the collapsed offset.
@@ -142,7 +147,7 @@ class FlattenContiguousRowMajorTransferReadPattern
 
   LogicalResult matchAndRewrite(vector::TransferReadOp transferReadOp,
                                 PatternRewriter &rewriter) const override {
-    auto loc = transferReadOp.getLoc();
+    Location loc = transferReadOp.getLoc();
     Value vector = transferReadOp.getVector();
     VectorType vectorType = cast<VectorType>(vector.getType());
     auto source = transferReadOp.getSource();
@@ -182,7 +187,7 @@ class FlattenContiguousRowMajorTransferReadPattern
     // 2.1. New dim exprs + affine map
     SmallVector<AffineExpr, 1> dimExprs{
         getAffineDimExpr(firstDimToCollapse, rewriter.getContext())};
-    auto collapsedMap =
+    AffineMap collapsedMap =
         AffineMap::get(collapsedRank, 0, dimExprs, rewriter.getContext());
 
     // 2.2 New indices
@@ -229,7 +234,7 @@ class FlattenContiguousRowMajorTransferWritePattern
 
   LogicalResult matchAndRewrite(vector::TransferWriteOp transferWriteOp,
                                 PatternRewriter &rewriter) const override {
-    auto loc = transferWriteOp.getLoc();
+    Location loc = transferWriteOp.getLoc();
     Value vector = transferWriteOp.getVector();
     VectorType vectorType = cast<VectorType>(vector.getType());
     Value source = transferWriteOp.getSource();
@@ -272,7 +277,7 @@ class FlattenContiguousRowMajorTransferWritePattern
     // 2.1. New dim exprs + affine map
     SmallVector<AffineExpr, 1> dimExprs{
         getAffineDimExpr(firstDimToCollapse, rewriter.getContext())};
-    auto collapsedMap =
+    AffineMap collapsedMap =
         AffineMap::get(collapsedRank, 0, dimExprs, rewriter.getContext());
 
     // 2.2 New indices
@@ -360,6 +365,221 @@ struct CanonicalizeTrivialReadAccessSubviewOpPattern
     rewriter.replaceOpWithNewOp<vector::TransferReadOp>(
         readOp, readOp.getType(), subViewOp.getSource(), newIndices,
         readOp.getPadding(), readOp.getInBoundsValues());
+    return success();
+  }
+};
+
+/// `vector.transfer_write` gets converted to `llvm.store` in the pass
+/// --convert-vector-to-llvm. The `llvm.store` persists through the
+/// translation to LLVMIR, and then through peano's `opt`. Then in peano's
+/// `llc` (VLIW code generation), what the `llvm.store` gets converted to
+/// depends on the size of the store. This pattern splits a transfer_read
+/// into smaller transfer_reads based on the best size for hitting vector
+/// assembly loads after `llc`. For example, assuming 128 bytes is the
+/// best store size for vectorization.
+///
+/// INPUT:
+///  %c0 = arith.constant 0 : index
+///  %cst = arith.constant dense<0> : vector<256xi32>
+///  %a = memref.alloc() : memref<256xi32>
+///  vector.transfer_write %cst, %a[%c0] {..} vector<256xi32>, memref<256xi32>
+///
+/// OUTPUT:
+///  %c0 = arith.constant 0 : index
+///  %c32 = arith.constant 32 : index
+///  %c256 = arith.constant 256 : index
+///  %cst = arith.constant dense<0> : vector<32xi32>
+///  %a = memref.alloc() : memref<256xi32>
+///  scf.for %arg0 = %c0 to %c256 step %c32 {
+///    vector.transfer_write %cst, %a[%arg0] {..} : vector<32xi32>,
+///                                                 memref<256xi32>
+///  }
+///
+/// Note: the closest thing to this that I can find upstream is
+/// `populateVectorUnrollPatterns` but that isn't quite what we want.
+/// Itunrolls the loop (this pass doesn't) and creates ops to extract
+/// and insert strided slices that are not needed when the vector is a splat
+/// value.
+
+struct SerializeSplatTransferReadWithTargetLoadSize
+    : public OpRewritePattern<vector::TransferWriteOp> {
+  using OpRewritePattern<vector::TransferWriteOp>::OpRewritePattern;
+  SerializeSplatTransferReadWithTargetLoadSize(MLIRContext *context,
+                                               AMDAIE::AMDAIEDevice d)
+      : OpRewritePattern<vector::TransferWriteOp>(context),
+        deviceModel(AMDAIE::getDeviceModel(d)) {}
+
+ private:
+  AMDAIE::AMDAIEDeviceModel deviceModel;
+  LogicalResult matchAndRewrite(vector::TransferWriteOp writeOp,
+                                PatternRewriter &rewriter) const override {
+    Value writeDestination = writeOp.getSource();
+    MemRefType writeDestinationType =
+        dyn_cast<MemRefType>(writeDestination.getType());
+    {
+      assert(writeDestinationType && "transfer_write must write to memref");
+      if (!writeDestinationType.hasStaticShape()) {
+        return rewriter.notifyMatchFailure(writeOp, "source has dynamic shape");
+      }
+      MemRefLayoutAttrInterface layout = writeDestinationType.getLayout();
+      if (layout && !layout.isIdentity()) {
+        return rewriter.notifyMatchFailure(writeOp,
+                                           "source has non-identity layout");
+      }
+    }
+    // This pass only succeeds when the memref that is written to a rank-1.
+    // There is already a pattern for flattening memrefs that should
+    // be included with this pattern in a pattern set.
+    if (writeDestinationType.getRank() != 1) {
+      return rewriter.notifyMatchFailure(writeOp, "memref isn't rank-1");
+    }
+
+    Operation::operand_range initialWriteIndices = writeOp.getIndices();
+    assert(initialWriteIndices.size() == 1 && "write destination is rank-1");
+    Attribute initialIndexAttr;
+    matchPattern(initialWriteIndices[0], m_Constant(&initialIndexAttr));
+    if (!initialIndexAttr) {
+      return rewriter.notifyMatchFailure(writeOp, "index isn't constant");
+    }
+    int64_t initialIndex = cast<IntegerAttr>(initialIndexAttr).getInt();
+
+    VectorType initialVectorType =
+        dyn_cast<VectorType>(writeOp.getVector().getType());
+    // Sanity checks on the vector operand:
+    assert(initialVectorType && "vector must be of vector type");
+    assert(writeDestinationType.getElementType() ==
+               initialVectorType.getElementType() &&
+           "element types must match");
+
+    // Find the arith.constant that the vector operand is a view of, if it is
+    // one.
+    arith::ConstantOp constantVectorSource = [&writeOp]() -> arith::ConstantOp {
+      Value current = writeOp.getVector();
+      while (Operation *op = current.getDefiningOp()) {
+        if (auto cOp = dyn_cast<arith::ConstantOp>(op)) return cOp;
+        if (op->getNumOperands() != 1) return {};
+        current = op->getOperand(0);
+      }
+      return {};
+    }();
+    if (!constantVectorSource) {
+      return rewriter.notifyMatchFailure(
+          writeOp, "vector isn't derived from arith.constant");
+    }
+
+    // Get the splat value of the constant vector.
+    auto maybeSplat = [&]() -> FailureOr<Attribute> {
+      TypedAttr constantValue = constantVectorSource.getValue();
+      auto splatAttr = dyn_cast<SplatElementsAttr>(constantValue);
+      if (!splatAttr || !splatAttr.isSplat()) {
+        return rewriter.notifyMatchFailure(writeOp, "constant isn't a splat");
+      }
+      return splatAttr.getSplatValue<Attribute>();
+    }();
+    if (failed(maybeSplat)) return failure();
+    Attribute splat = maybeSplat.value();
+
+    int64_t bytesPerWrite = deviceModel.getPreferredLoadBytes();
+
+    Type elementType = writeDestinationType.getElementType();
+
+    int64_t nbWriteElements = initialVectorType.getNumElements();
+    int64_t bitsPerElement = elementType.getIntOrFloatBitWidth();
+    int64_t bytesPerElement = bitsPerElement / 8;
+    int64_t totalBytes = nbWriteElements * bytesPerElement;
+    int64_t elementsPerWrite = bytesPerWrite / bytesPerElement;
+    int64_t nbLoopWrites = nbWriteElements / elementsPerWrite;
+    int64_t nbTailBytes = totalBytes % bytesPerWrite;
+    int64_t nbTailElements = nbTailBytes / bytesPerElement;
+    int64_t tailStart = initialIndex + nbLoopWrites * elementsPerWrite;
+
+    if (totalBytes <= bytesPerWrite) {
+      return rewriter.notifyMatchFailure(writeOp,
+                                         "source is smaller than write size");
+    }
+
+    // Mae a transfer_write that writes to the original memref destination,
+    // but with an adjusted number of elements and an adjusted offset index.
+    auto makeTransferWrite = [&](uint32_t n, Value offset) {
+      VectorType type = VectorType::get({n}, elementType);
+      DenseElementsAttr attr = DenseElementsAttr::get(type, splat);
+      auto newConstantOp = rewriter.create<arith::ConstantOp>(
+          constantVectorSource.getLoc(), type, attr);
+      rewriter.create<vector::TransferWriteOp>(
+          writeOp.getLoc(), newConstantOp.getResult(), writeDestination, offset,
+          writeOp.getPermutationMapAttr(), writeOp.getInBoundsAttr());
+    };
+
+    auto getConstant = [&](int64_t v) {
+      return rewriter.create<arith::ConstantIndexOp>(writeOp.getLoc(), v);
+    };
+    // Construct the loop body.
+    rewriter.setInsertionPointAfter(writeOp);
+    Value cStart = getConstant(initialIndex);
+    Value cEnd = getConstant(tailStart);
+    Value cStep = getConstant(elementsPerWrite);
+    scf::ForOp loopOp =
+        rewriter.create<scf::ForOp>(writeOp.getLoc(), cStart, cEnd, cStep);
+
+    // TODO(newling) consider partial unrolling if beneficial, or some smarter
+    // way of deciding whether to attach no unrolling. Perhaps this should even
+    // be a separate pass. Preventing this from unrolling here for now, as I see
+    // it saves 1K PM bytes for a linalg.fill for a matmul.
+    mlir::LLVM::LoopUnrollAttr unrollAttr;
+    mlir::LLVM::LoopAnnotationAttr loopAnnotationAttr;
+    BoolAttr disableAttr = rewriter.getBoolAttr(true);
+    unrollAttr = mlir::LLVM::LoopUnrollAttr::get(
+        rewriter.getContext(), /*disable=*/disableAttr, /*count=*/{},
+        /*runtimeDisable=*/{}, /*full=*/{}, /*followupUnrolled=*/{},
+        /*followupRemainder=*/{}, /*followupAll=*/{});
+
+    loopAnnotationAttr = mlir::LLVM::LoopAnnotationAttr::get(
+        rewriter.getContext(), /*disableNonforced=*/{},
+        /*vectorize=*/{}, /*interleave=*/{}, /*unroll=*/unrollAttr,
+        /*unrollAndJam=*/{}, /*licm=*/{}, /*distribute=*/{},
+        /*pipeline=*/{},
+        /*peeled=*/{}, /*unswitch=*/{}, /*mustProgress=*/{},
+        /*isVectorized=*/{}, /*startLoc=*/{}, /*endLoc=*/{},
+        /*parallelAccesses=*/{});
+
+    // Add the llvm.loop_annotation attribute to the loop.
+    loopOp->setAttr("loop_annotation", loopAnnotationAttr);
+
+    // Create transfer_write inside loop body.
+    rewriter.setInsertionPointToStart(loopOp.getBody());
+    makeTransferWrite(elementsPerWrite, loopOp.getInductionVar());
+
+    // If there are tail elements, create a transfer_write for them.
+    if (nbTailElements > 0) {
+      rewriter.setInsertionPointAfter(loopOp);
+      makeTransferWrite(nbTailElements, cEnd);
+    }
+
+    // erase the original transfer_write
+    rewriter.eraseOp(writeOp);
+    return success();
+  }
+};
+
+/// TODO(newling) upstream this to MLIR.
+///
+/// A simple folder:
+/// ```
+/// %1 = memref.reinterpret_cast %0 ...
+/// %2 = memref.collapse_shape %1 ...
+/// ```
+/// If type of %2 is same as type of %0, replace uses of %2 with %0.
+struct FoldReinterpretCastFollowedByCollapseShapePattern
+    : public OpRewritePattern<memref::CollapseShapeOp> {
+  using OpRewritePattern<memref::CollapseShapeOp>::OpRewritePattern;
+  LogicalResult matchAndRewrite(memref::CollapseShapeOp collapseOp,
+                                PatternRewriter &rewriter) const override {
+    auto reinterpretOp = dyn_cast_if_present<memref::ReinterpretCastOp>(
+        collapseOp.getOperand().getDefiningOp());
+    if (!reinterpretOp) return failure();
+    auto reinterpretInputType = reinterpretOp.getSource().getType();
+    if (reinterpretInputType != collapseOp.getType()) return failure();
+    rewriter.replaceOp(collapseOp, reinterpretOp.getSource());
     return success();
   }
 };
@@ -663,9 +883,9 @@ getUnsqueezedShapeAndInBounds(uint64_t inputRank,
   return {newShape, newInBounds};
 }
 
-/// This pattern rewrites a `vector.transfer_write` with a non minor identity
-/// permutation map, with a minor-identity permutation map, if possible. For
-/// example,
+/// This pattern rewrites a `vector.transfer_write` without a minor-identity
+/// permutation map as one with a minor-identity permutation map, if possible.
+/// For example,
 ///
 /// ```
 /// #map = affine_map<(d0, d1, d2, d3) -> (d1, d3)>
@@ -1031,13 +1251,23 @@ struct CanonicalizeVectorForAIEVecPass
   }
 
   void getDependentDialects(DialectRegistry &registry) const override {
-    registry.insert<arith::ArithDialect, memref::MemRefDialect,
-                    vector::VectorDialect, affine::AffineDialect>();
+    registry.insert<affine::AffineDialect, arith::ArithDialect,
+                    memref::MemRefDialect, scf::SCFDialect,
+                    vector::VectorDialect, LLVM::LLVMDialect>();
   }
 
   void runOnOperation() override {
     auto op = getOperation();
     MLIRContext *context = &getContext();
+
+    std::optional<AMDAIE::AMDAIEDevice> maybeDevice =
+        AMDAIE::getConfigAMDAIEDeviceFromAncestor(op);
+    if (!maybeDevice.has_value()) {
+      op->emitOpError(
+          "doesn't have target_device specified in a parent module.");
+      return signalPassFailure();
+    }
+    AMDAIE::AMDAIEDevice device = maybeDevice.value();
 
     {
       RewritePatternSet patterns(context);
@@ -1078,6 +1308,9 @@ struct CanonicalizeVectorForAIEVecPass
       // cycles.
       RewritePatternSet patterns(context);
       populateBubbleSignExtensionsLate(patterns);
+      patterns.add<FoldReinterpretCastFollowedByCollapseShapePattern>(context);
+      patterns.add<SerializeSplatTransferReadWithTargetLoadSize>(context,
+                                                                 device);
       (void)applyPatternsGreedily(op, std::move(patterns));
     }
   }
@@ -1236,9 +1469,9 @@ struct AlignTransferReadsPass
   }
 
   void getDependentDialects(DialectRegistry &registry) const override {
-    registry
-        .insert<arith::ArithDialect, memref::MemRefDialect,
-                vector::VectorDialect, affine::AffineDialect, AIEVecDialect>();
+    registry.insert<affine::AffineDialect, AIEVecDialect, arith::ArithDialect,
+                    LLVM::LLVMDialect, memref::MemRefDialect,
+                    vector::VectorDialect>();
   }
 
   void runOnOperation() override {
@@ -1280,8 +1513,9 @@ struct DetectNonCanonicalOpsPass
   }
 
   void getDependentDialects(DialectRegistry &registry) const override {
-    registry.insert<arith::ArithDialect, memref::MemRefDialect,
-                    vector::VectorDialect, affine::AffineDialect>();
+    registry
+        .insert<affine::AffineDialect, arith::ArithDialect, LLVM::LLVMDialect,
+                memref::MemRefDialect, vector::VectorDialect>();
   }
 
   void runOnOperation() override {